Add Therapist Filtering and Rename Use Case Folder
